feat(auction): add third auction item winner payment reminder email
