+++ backend/utils/campaign-utils/createAuctionItemDocument.js
@@ -1,0 +1,71 @@
+import { AuctionItem, AuctionItemPhoto } from '../../models/campaignModel.js';
+import Error from '../../models/errorModel.js';
+import { logEvent } from '../logHelpers.js';
+
+async function createAuctionItemDocument(log, data) {
+  logEvent(log, 'INITIATE CREATE AUCTION ITEM', data);
+
+  try {
+    const auctionItemPhotos = await Promise.all(
+        data.photos.map(async (photo) => {
+          const auctionItemPhoto = new AuctionItemPhoto({
+            name: photo.name,
+            url: photo.url,
+            size: photo.size,
+          });
+  
+          const savedPhoto = await auctionItemPhoto.save();
+          return savedPhoto; // Return the saved photo document
+        })
+      );
+
+    const auctionItemData = {
+        auction: data.auction,
+        name: data.name,
+        description: data.description,
+        photos: auctionItemPhotos.map((photo) => photo._id),
+        requiresShipping: data.requiresShipping,
+        shippingCosts: data.shippingCosts,
+        totalQuantity: data.totalQuantity || 1, // default to 1 for auction items
+      };
+
+    let auctionItem;
+    if (data.sellingFormat === 'fixed') {
+      auctionItem = new AuctionItem({
+        ...auctionItemData,
+        sellingFormat: 'fixed',
+        buyNowPrice: data.buyNowPrice,
+        itemBtnText: 'Buy Now',
+        isFixed: true,
+      });
+      logEvent(log, 'CREATED FIXED AUCTION ITEM', auctionItem);
+    } else {
+      auctionItem = new AuctionItem({
+        ...auctionItemData,
+        sellingFormat: 'auction',
+        startingPrice: data.startingPrice,
+        currentBid: data.startingPrice,
+        totalBids: 0,
+        itemBtnText: 'Place Bid',
+        minimumBid: data.minimumBid,
+        isAuction: true,
+      });
+      logEvent(log, 'CREATED AUCTION ITEM', auctionItem);
+    }
+
+   const savedAuctionItem = await auctionItem.save();
+   logEvent(log, 'AUCTION ITEM SAVED', savedAuctionItem);
+
+   return savedAuctionItem;
+  } catch (err) {
+    logEvent(log, 'ERROR CREATING AUCTION ITEM', err);
+
+    await Error.create({
+      functionName: 'CREATE_AUCTION_ITEM_DOCUMENT_FUCNTION_PRIVATE_ADMIN',
+      name: err.name,
+      message: err.message,
+    });
+  }
+}
+
+export default createAuctionItemDocument;
