+++ backend/utils/emails/outBidNotification.js
@@ -1,6 +1,10 @@
+import { Bid } from '../../models/campaignModel.js';
 import Error from '../../models/errorModel.js';
+import { logEvent, prepareLog } from '../logHelpers.js';
 
 const outBidNotification = async (pugEmail, data) => {
+  const log = await prepareLog('INITIATE OUT BID NOTIFICATION EMAIL');
+  logEvent(log, 'OUT BID NOTIFICATION EMAIL', data);
   await pugEmail
     .send({
       template: 'outbidnotification',
@@ -16,8 +20,9 @@ const outBidNotification = async (pugEmail, data) => {
         link: data.link,
       },
     })
-    .then(() => {
-      console.log('OUT BID NOTIFICATION SUCCESS');
+    .then(async () => {
+      await Bid.findByIdAndUpdate(data.previousTopBidId, { outBidEmailSent: true }, { new: true });
+      logEvent(log, `SUCCESSFULLY SENT OUT BID EMAIL TO ${data.email}`);
     })
     .catch(async (err) => {
       await Error.create({
