+++ emails/forgotpassword/html.pug
@@ -1,0 +1,263 @@
+doctype transitional
+html(xmlns='http://www.w3.org/1999/xhtml')
+  head
+    meta(name='viewport' content='width=device-width')
+    meta(http-equiv='Content-Type' content='text/html; charset=UTF-8')
+    title forgot-password
+    style(type='text/css').
+      * {
+        font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+      }
+      body {
+        -webkit-font-smoothing: antialiased;
+        -webkit-text-size-adjust: none;
+        align-items: center;
+        height: 100%;
+        background-color: #f6f6f6;
+        padding: 24px;
+      }
+
+      .container {
+        display: block !important;
+        max-width: 600px !important;
+        margin: 0 auto !important;
+        /* makes it centered */
+        clear: both !important;
+        background-color: #fff;
+        padding: 24px;
+      }
+
+      .content {
+        max-width: 800px;
+        margin: 0 auto;
+        display: block;
+        padding: 36px;
+      }
+
+      .footer {
+        height: 20px;
+      }
+
+      .header-section {
+        margin-bottom: 24px;
+      }
+
+      h3 {
+        margin-top: 0;
+        margin-bottom: 8px;
+      }
+
+      .secondary-font {
+        color: #dcdde1;
+        margin-bottom: 32px;
+        font-size: 14px;
+      }
+
+      .horizontal-line {
+        border-bottom: 1px solid #f2f2f2;
+
+      }
+
+      table {
+        padding: 24px 0;
+      }
+
+      .table-header {
+        margin-bottom: 8px;
+      }
+
+      div.auction-item {
+        display: flex;
+        justify-content: between;
+        align-items: center;
+        max-width: 800px;
+        width: 100%;
+        margin-top: 48px;
+        margin-bottom: 32px;
+      }
+
+      .picture-name-desc {
+        float: left;
+        width: 100%;
+      }
+
+      .total-container {
+        width: 100%;
+        display: block;
+        margin-bottom: 32px;
+      }
+
+      .total-block {
+        float: right;
+        max-width: 250px;
+        width: 100%;
+        margin: 32px 0;
+      }
+
+      .flex {
+        display: flex;
+      }
+
+      .total {
+        display: flex;
+        font-weight: 400;
+        font-size: 16px;
+        padding: 8px 0;
+      }
+
+      .clear {
+        clear: both;
+        margin-bottom: 48px;
+      }
+
+      .thank-you {
+        color: #6c6f79;
+        font-size: 18px;
+        font-weight: 400;
+      }
+
+      .lpdr {
+        font-weight: 400;
+      }
+
+      .aligncenter {
+        text-align: center;
+      }
+
+      .float-right {
+        font-weight: 400;
+        float: right;
+        width: fit-content;
+      }
+
+      .float-left {
+        float: left;
+        width: 75%;
+      }
+
+      .total-price {
+        float: right;
+        width: fit-content;
+      }
+
+      .footer {
+        margin-top: 48px;
+        background: #fcfbfe;
+        padding: 20px 36px;
+        max-width: 800px;
+        height: 20px;
+      }
+
+      .help {
+        text-align: center !important;
+      }
+
+      .rescue {
+        float: right;
+      }
+      span {
+        color: #9761aa;
+      }
+
+      .logo {
+        width:85px;
+        height:auto;
+        object-fit: cover;
+        margin-bottom: 24px;
+      }
+
+      div {
+        font-size: 11px;
+      }
+
+      .name {
+        font-size: 16px;
+      }
+      .desc, .details {
+        font-size: 14px;
+      }
+
+      .details {
+        width: fit-content;
+        float: left;
+        font-size: 14px;
+      }
+      .details-data {
+        width: fit-content;
+        float: right;
+        font-size: 14px;
+      }
+      .total {
+        font-weight: 600;
+      }
+
+      .price {
+        font-size: 14px;
+        width: fit-content;
+        float: right;
+      }
+
+      .congrats {
+        font-size: 28px;
+        text-align: center;
+        font-weight: 600;
+      }
+
+      .pay-items-link {
+        background: #22c2b7;
+        padding: 8px 44px;
+        color: white;
+        font-size: 16px;
+        font-weight: 600;
+        margin: 0 auto 24px;
+        border-radius: 6px;
+      }
+
+      .text {
+        font-size: 14px;
+        margin: 0 auto;
+      }
+
+      .align-center {
+        width:100%;
+        margin-inline: auto;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        text-align: center;
+      }
+
+      @media only screen and (max-width: 640px) {
+        body {
+          padding: 0 !important;
+        }
+        h3 {
+          font-size: 16px !important;
+        }
+        .container {
+          padding: 0 !important;
+          width: 100% !important;
+        }
+        .content {
+          padding: 0 !important;
+        }
+        .footer {
+          margin-top: 32px !important;
+          padding: 20px 0 !important;
+          background: #fff !important;
+        }
+      }
+
+  body(itemscope='' itemtype='http://schema.org/EmailMessage')
+    .container(width='600px')
+      .header-section
+        img.logo(src='https://i.ibb.co/GMWQxTM/logo.png' alt='little paws dachshund rescue')
+        p.congrats Reset Your Password
+      .horizontal-line
+      p You are receiving this because you (or someone else) have requested the reset of the password for your account. If you did not request this, please ignore this email and your password will remain unchanged
+      div.align-center
+        a.pay-items-link(href=`https://www.littlepawsdr.org/auth/reset-password/${token}`) Reset Password
+      .footer
+        -var contactLink = `<a href='https://www.littlepawsdr.org/contact-us')>Contact</a>`;
+        .help Need Help? Visit our <span>!{contactLink}</span> page.
+
