+++ emails/forgotpassword/subject.pug
@@ -1,0 +1,1 @@
+= `Reset Password`
