+++ emails/forgotpassword/text.pug
@@ -1,0 +1,1 @@
+= `Hello, ${email}. Please check another browser to receive email order confirmation. Just in case, here is your order id: ${id}`
