+++ frontend/src/components/data/campaign/auction-notification-data.ts
@@ -1,0 +1,20 @@
+const auctionNotificationsData = [
+  {
+    textKey: 'Out Bid',
+    freq: 'Immediate',
+  },
+  {
+    textKey: 'Winning Bidder Confirmation',
+    freq: 'Immediate',
+  },
+  {
+    textKey: 'First Payment Reminder',
+    freq: '2 Days',
+  },
+  {
+    textKey: 'Second Payment Reminder',
+    freq: '4 Days',
+  },
+];
+
+export default auctionNotificationsData;
