+++ frontend/src/routes/Admin/Campaigns/Auction/Notifications.tsx
@@ -1,3 +1,5 @@
+import auctionNotificationsData from '../../../../components/data/campaign/auction-notification-data';
+
 const Notifications = () => {
   return (
     <div className='bg-white border border-gray-300 rounded-md w-full'>
@@ -12,36 +14,24 @@ const Notifications = () => {
           Email
         </div>
       </div>
-      <div className='grid grid-cols-6 px-8 py-[20px] border-b border-slate-200'>
-        <div className='col-span-4'>
-          <div className='flex items-center space-x-2 font-Matter-Regular'>Auction end</div>
+      {auctionNotificationsData.map((obj: any, i) => (
+        <div
+          key={i}
+          className={`grid grid-cols-6 px-8 py-[20px] ${
+            i !== 3 ? ' border-b border-slate-200' : ''
+          }`}
+        >
+          <div className='col-span-4'>
+            <div className='flex items-center space-x-2 font-Matter-Regular'>{obj.textKey}</div>
+          </div>
+          <div className='col-span-1 col-start-5 place-self-center font-Matter-Light text-sm'>
+            {obj.freq}
+          </div>
+          <div className='col-span-1 col-start-6 place-self-center'>
+            <i className='fas fa-check'></i>
+          </div>
         </div>
-        <div className='col-span-1 col-start-6 place-self-center'>
-          <i className='fas fa-check'></i>
-        </div>
-      </div>
-      <div className='grid grid-cols-6 px-8 py-[20px] border-b border-slate-200'>
-        <div className='col-span-4'>
-          <div className='flex items-center space-x-2 font-Matter-Regular'>Outstanding invoice</div>
-        </div>
-        <div className='col-span-1 col-start-5 place-self-center font-Matter-Light text-sm'>
-          2 Days
-        </div>
-        <div className='col-span-1 col-start-6 place-self-center'>
-          <i className='fas fa-check'></i>
-        </div>
-      </div>
-      <div className='grid grid-cols-6 px-8 py-[20px]'>
-        <div className='col-span-4'>
-          <div className='flex items-center space-x-2 font-Matter-Regular'>Out Bid</div>
-        </div>
-        <div className='col-span-1 col-start-5 place-self-center font-Matter-Light text-sm'>
-          Immediate
-        </div>
-        <div className='col-span-1 col-start-6 place-self-center'>
-          <i className='fas fa-check'></i>
-        </div>
-      </div>
+      ))}
     </div>
   );
 };
